feat: modularize portfolio sections with new responsive background and UI components
